+++ server/subsonic/helpers.go
@@ -85,8 +85,8 @@ func toArtist(r *http.Request, a model.Artist) responses.Artist {
 	artist := responses.Artist{
 		Id:             a.ID,
 		Name:           a.Name,
-		AlbumCount:     a.AlbumCount,
-		UserRating:     a.Rating,
+		AlbumCount:     int32(a.AlbumCount),
+		UserRating:     int32(a.Rating),
 		CoverArt:       a.CoverArtID().String(),
 		ArtistImageUrl: public.ImageURL(r, a.CoverArtID(), 600),
 	}
